[op-creds] Fix RemoveFabric command
RemoveFabric command called for the fabric that is used to
process the message crashes in SessionHolder. The reason is
that sessions for the removed fabric are expired when the
command's exchange is closed after sending the response
and that triggers closing all associated session holders.
An exchange's session holder closes the exchange which
in this case results in an attempt to close an exchange
that is in the middle of closing...